Added the ability to read YAML config files for both master-config.yaml and the existing cli variables.
A YAML configuration file can be specified with `-config=/path/to/file.yaml`.
If -config is specified, any values in the file will clobber ones provided on
the command line, which is the same behavior that openshift exhibits.

Added unit tests for both configuration types, and yaml files that are used by
those tests.