Apply "Allow IP" and "Allow Non-IP" ACLs by default
Also fix existing tests and add new ones for CreateZone() and DeleteZone()

I changed some function signatures because the parameters were no longer
necessary in the zone-per-project model and removed functions dealing with zone
templates.  This change broke some existing tests, but those tests have been
updated to match the current set of functions.  GetZoneID() was also updated to
act like GetDomainID() and GetSubnetID(), in that when no zone was found with a
name that matched the input, it will return an error with the message 'Zone not
found', rather than an 'unexpected error.'