Fix VRS-4035
Cases where there was a conflict with ACL template priority weren't handled
properly.  This patch retries with successively higher priority numbers
(technically, lower priority) until there is one without an issue, unless the
error is that there is a conflict with the name of the template, which indicates
that the template we want to use already exists.

Also fixes a bug I found as while reproducing this one, where if a default
subnet for a specific project already existed, the ACL entries for that project
wouldn't be created.