Add user to existing admin group, instead of clobbering the list
The old method of adding the admin user to the `ORGADMIN` group involved
`PUT`-ing just the new user to `/groups/<admin-group-id>/users` because it was
believed that `PUT` against that endpoint would act as a diff, adding the user
to the existing list.  It was found that that was not true, and instead, the
list provided in the `PUT` call's body replaces the existing one.  This patch
uses `GET` calls to get all the existing users in `ORGADMIN`, adds the new user
to the list, then does a `PUT` of the full list, resulting in adding the new
user to `ORGADMIN`.

Reorder arguments to `AddNetworkMacroToNMG` to match the function name

Function name is `AddNetworkMacroToNMG` (`NetworkMacro`, then `NMG`), but the
arguments were in the order `NMG`, then `NetworkMacro`.  This patch changes the
arguments to be in the same order as the nouns in the function name, and changes
the order of the arguments where it's called.